chore(unit-test): create tests for SvgPath class
