+++ shared/src/commonTest/kotlin/dev/tonholo/s2c/domain/svg/SvgPathCurveToTests.kt
@@ -1,9 +1,7 @@
-package dev.tonholo.s2c.domain
+package dev.tonholo.s2c.domain.svg
 
-import dev.tonholo.s2c.domain.svg.SvgElementNode
-import dev.tonholo.s2c.domain.svg.SvgPathNode
-import dev.tonholo.s2c.domain.svg.asNode
-import dev.tonholo.s2c.domain.xml.XmlRootNode
+import dev.tonholo.s2c.domain.ImageVectorNode
+import dev.tonholo.s2c.domain.PathNodes
 import dev.tonholo.s2c.extensions.removeTrailingZero
 import kotlin.test.Test
 import kotlin.test.assertContains
@@ -12,28 +10,24 @@ import kotlin.test.assertFalse
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
-class PathNodesReflectiveCurveToTests {
-    data class ReflectiveCurveToParams(
+class SvgPathCurveToTests : BaseSvgTest() {
+    data class CurveToParams(
         val x1: Float,
         val y1: Float,
         val x2: Float,
         val y2: Float,
+        val x3: Float,
+        val y3: Float,
     ) {
-        override fun toString(): String = "$x1 $y1 $x2 $y2".removeTrailingZero()
+        override fun toString(): String = "$x1 $y1 $x2 $y2 $x3 $y3".removeTrailingZero()
     }
 
-    private val root = SvgElementNode(
-        parent = XmlRootNode(children = mutableSetOf()),
-        children = mutableSetOf(),
-        attributes = mutableMapOf(),
-    )
-
     @Test
-    fun `ensure a 's' command is parsed to ReflectiveCurveTo relative`() {
+    fun `ensure a 'c' command is parsed to CurveTo relative`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "s2,-2 4,5"),
+            attributes = mutableMapOf("d" to "c2,8 8,8 8,5"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -41,17 +35,17 @@ class PathNodesReflectiveCurveToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.ReflectiveCurveTo>(this)
+            assertIs<PathNodes.CurveTo>(this)
             assertTrue(isRelative)
         }
     }
 
     @Test
-    fun `ensure a 'S' command is parsed to ReflectiveCurveTo`() {
+    fun `ensure a 'C' command is parsed to CurveTo`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "S2,-2 4,5"),
+            attributes = mutableMapOf("d" to "C2,8 8,8 8,5"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -59,29 +53,33 @@ class PathNodesReflectiveCurveToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.ReflectiveCurveTo>(this)
+            assertIs<PathNodes.CurveTo>(this)
             assertFalse(isRelative)
         }
     }
 
     @Test
-    fun `ensure materialize generates reflectiveCurveTo instruction properly`() {
+    fun `ensure materialize generates curveTo instruction properly`() {
         // Arrange
-        val nonRelative = ReflectiveCurveToParams(
+        val nonRelative = CurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
+            x3 = 0f,
+            y3 = 123f,
         )
-        val relative = ReflectiveCurveToParams(
+        val relative = CurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
+            x3 = 10f,
+            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "S$nonRelative s$relative"),
+            attributes = mutableMapOf("d" to "C$nonRelative c$relative"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -92,12 +90,14 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// S $this
-                |reflectiveCurveTo(
+                |// C $this
+                |curveTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
                 |    x2 = ${x2}f,
                 |    y2 = ${y2}f,
+                |    x3 = ${x3}f,
+                |    y3 = ${y3}f,
                 |)
                 |""".trimMargin()
             )
@@ -106,12 +106,14 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// s $this
-                |reflectiveCurveToRelative(
+                |// c $this
+                |curveToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
                 |    dx2 = ${x2}f,
                 |    dy2 = ${y2}f,
+                |    dx3 = ${x3}f,
+                |    dy3 = ${y3}f,
                 |)
                 |""".trimMargin()
             )
@@ -119,23 +121,27 @@ class PathNodesReflectiveCurveToTests {
     }
 
     @Test
-    fun `ensure materialize generates reflectiveCurveTo with a close instruction properly`() {
+    fun `ensure materialize generates curveTo with a close instruction properly`() {
         // Arrange
-        val nonRelative = ReflectiveCurveToParams(
+        val nonRelative = CurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
+            x3 = 0f,
+            y3 = 123f,
         )
-        val relative = ReflectiveCurveToParams(
+        val relative = CurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
+            x3 = 10f,
+            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "S${nonRelative}z s${relative}z"),
+            attributes = mutableMapOf("d" to "C${nonRelative}z c${relative}z"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -146,12 +152,14 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// S ${this}z
-                |reflectiveCurveTo(
+                |// C ${this}z
+                |curveTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
                 |    x2 = ${x2}f,
                 |    y2 = ${y2}f,
+                |    x3 = ${x3}f,
+                |    y3 = ${y3}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -161,12 +169,14 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// s ${this}z
-                |reflectiveCurveToRelative(
+                |// c ${this}z
+                |curveToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
                 |    dx2 = ${x2}f,
                 |    dy2 = ${y2}f,
+                |    dx3 = ${x3}f,
+                |    dy3 = ${y3}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -175,23 +185,27 @@ class PathNodesReflectiveCurveToTests {
     }
 
     @Test
-    fun `ensure materialize generates reflectiveCurveTo with inlined parameters and no comment`() {
+    fun `ensure materialize generates curveTo with inlined parameters and no comment when minified`() {
         // Arrange
-        val nonRelative = ReflectiveCurveToParams(
+        val nonRelative = CurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
+            x3 = 0f,
+            y3 = 123f,
         )
-        val relative = ReflectiveCurveToParams(
+        val relative = CurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
+            x3 = 10f,
+            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "S$nonRelative s$relative"),
+            attributes = mutableMapOf("d" to "C$nonRelative c$relative"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -202,48 +216,56 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("reflectiveCurveTo(")
+                    append("curveTo(")
                     append("x1 = ${x1}f,")
                     append(" y1 = ${y1}f,")
                     append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f")
+                    append(" y2 = ${y2}f,")
+                    append(" x3 = ${x3}f,")
+                    append(" y3 = ${y3}f")
                     append(")")
-                },
+                }
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("reflectiveCurveToRelative(")
+                    append("curveToRelative(")
                     append("dx1 = ${x1}f,")
                     append(" dy1 = ${y1}f,")
                     append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f")
+                    append(" dy2 = ${y2}f,")
+                    append(" dx3 = ${x3}f,")
+                    append(" dy3 = ${y3}f")
                     append(")")
-                },
+                }
             )
         }
     }
 
     @Test
-    fun `ensure materialize generates reflectiveCurveTo with inlined parameters and no comment and close instruction when minified`() {
+    fun `ensure materialize generates curveTo with inlined parameters and no comment with close instruction when minified`() {
         // Arrange
-        val nonRelative = ReflectiveCurveToParams(
+        val nonRelative = CurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
+            x3 = 0f,
+            y3 = 123f,
         )
-        val relative = ReflectiveCurveToParams(
+        val relative = CurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
+            x3 = 10f,
+            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "S${nonRelative}z s${relative}z"),
+            attributes = mutableMapOf("d" to "C${nonRelative}z c${relative}z"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -254,30 +276,34 @@ class PathNodesReflectiveCurveToTests {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("reflectiveCurveTo(")
+                    append("curveTo(")
                     append("x1 = ${x1}f,")
                     append(" y1 = ${y1}f,")
                     append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f")
+                    append(" y2 = ${y2}f,")
+                    append(" x3 = ${x3}f,")
+                    append(" y3 = ${y3}f")
                     append(")")
                     appendLine()
                     append("close()")
-                },
+                }
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("reflectiveCurveToRelative(")
+                    append("curveToRelative(")
                     append("dx1 = ${x1}f,")
                     append(" dy1 = ${y1}f,")
                     append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f")
+                    append(" dy2 = ${y2}f,")
+                    append(" dx3 = ${x3}f,")
+                    append(" dy3 = ${y3}f")
                     append(")")
                     appendLine()
                     append("close()")
-                },
+                }
             )
         }
     }
