+++ shared/src/commonTest/kotlin/dev/tonholo/s2c/domain/svg/SvgPathReflectiveCurveToTests.kt
@@ -1,9 +1,7 @@
-package dev.tonholo.s2c.domain
+package dev.tonholo.s2c.domain.svg
 
-import dev.tonholo.s2c.domain.svg.SvgElementNode
-import dev.tonholo.s2c.domain.svg.SvgPathNode
-import dev.tonholo.s2c.domain.svg.asNode
-import dev.tonholo.s2c.domain.xml.XmlRootNode
+import dev.tonholo.s2c.domain.ImageVectorNode
+import dev.tonholo.s2c.domain.PathNodes
 import dev.tonholo.s2c.extensions.removeTrailingZero
 import kotlin.test.Test
 import kotlin.test.assertContains
@@ -12,30 +10,22 @@ import kotlin.test.assertFalse
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
-class PathNodesCurveToTests {
-    data class CurveToParams(
+class SvgPathReflectiveCurveToTests : BaseSvgTest() {
+    data class ReflectiveCurveToParams(
         val x1: Float,
         val y1: Float,
         val x2: Float,
         val y2: Float,
-        val x3: Float,
-        val y3: Float,
     ) {
-        override fun toString(): String = "$x1 $y1 $x2 $y2 $x3 $y3".removeTrailingZero()
+        override fun toString(): String = "$x1 $y1 $x2 $y2".removeTrailingZero()
     }
 
-    private val root = SvgElementNode(
-        parent = XmlRootNode(children = mutableSetOf()),
-        children = mutableSetOf(),
-        attributes = mutableMapOf(),
-    )
-
     @Test
-    fun `ensure a 'c' command is parsed to CurveTo relative`() {
+    fun `ensure a 's' command is parsed to ReflectiveCurveTo relative`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "c2,8 8,8 8,5"),
+            attributes = mutableMapOf("d" to "s2,-2 4,5"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -43,17 +33,17 @@ class PathNodesCurveToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.CurveTo>(this)
+            assertIs<PathNodes.ReflectiveCurveTo>(this)
             assertTrue(isRelative)
         }
     }
 
     @Test
-    fun `ensure a 'C' command is parsed to CurveTo`() {
+    fun `ensure a 'S' command is parsed to ReflectiveCurveTo`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "C2,8 8,8 8,5"),
+            attributes = mutableMapOf("d" to "S2,-2 4,5"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -61,33 +51,29 @@ class PathNodesCurveToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.CurveTo>(this)
+            assertIs<PathNodes.ReflectiveCurveTo>(this)
             assertFalse(isRelative)
         }
     }
 
     @Test
-    fun `ensure materialize generates curveTo instruction properly`() {
+    fun `ensure materialize generates reflectiveCurveTo instruction properly`() {
         // Arrange
-        val nonRelative = CurveToParams(
+        val nonRelative = ReflectiveCurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
-            x3 = 0f,
-            y3 = 123f,
         )
-        val relative = CurveToParams(
+        val relative = ReflectiveCurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
-            x3 = 10f,
-            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "C$nonRelative c$relative"),
+            attributes = mutableMapOf("d" to "S$nonRelative s$relative"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -98,14 +84,12 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// C $this
-                |curveTo(
+                |// S $this
+                |reflectiveCurveTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
                 |    x2 = ${x2}f,
                 |    y2 = ${y2}f,
-                |    x3 = ${x3}f,
-                |    y3 = ${y3}f,
                 |)
                 |""".trimMargin()
             )
@@ -114,14 +98,12 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// c $this
-                |curveToRelative(
+                |// s $this
+                |reflectiveCurveToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
                 |    dx2 = ${x2}f,
                 |    dy2 = ${y2}f,
-                |    dx3 = ${x3}f,
-                |    dy3 = ${y3}f,
                 |)
                 |""".trimMargin()
             )
@@ -129,27 +111,23 @@ class PathNodesCurveToTests {
     }
 
     @Test
-    fun `ensure materialize generates curveTo with a close instruction properly`() {
+    fun `ensure materialize generates reflectiveCurveTo with a close instruction properly`() {
         // Arrange
-        val nonRelative = CurveToParams(
+        val nonRelative = ReflectiveCurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
-            x3 = 0f,
-            y3 = 123f,
         )
-        val relative = CurveToParams(
+        val relative = ReflectiveCurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
-            x3 = 10f,
-            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "C${nonRelative}z c${relative}z"),
+            attributes = mutableMapOf("d" to "S${nonRelative}z s${relative}z"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -160,14 +138,12 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// C ${this}z
-                |curveTo(
+                |// S ${this}z
+                |reflectiveCurveTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
                 |    x2 = ${x2}f,
                 |    y2 = ${y2}f,
-                |    x3 = ${x3}f,
-                |    y3 = ${y3}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -177,14 +153,12 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// c ${this}z
-                |curveToRelative(
+                |// s ${this}z
+                |reflectiveCurveToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
                 |    dx2 = ${x2}f,
                 |    dy2 = ${y2}f,
-                |    dx3 = ${x3}f,
-                |    dy3 = ${y3}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -193,27 +167,23 @@ class PathNodesCurveToTests {
     }
 
     @Test
-    fun `ensure materialize generates curveTo with inlined parameters and no comment when minified`() {
+    fun `ensure materialize generates reflectiveCurveTo with inlined parameters and no comment`() {
         // Arrange
-        val nonRelative = CurveToParams(
+        val nonRelative = ReflectiveCurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
-            x3 = 0f,
-            y3 = 123f,
         )
-        val relative = CurveToParams(
+        val relative = ReflectiveCurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
-            x3 = 10f,
-            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "C$nonRelative c$relative"),
+            attributes = mutableMapOf("d" to "S$nonRelative s$relative"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -224,56 +194,48 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("curveTo(")
+                    append("reflectiveCurveTo(")
                     append("x1 = ${x1}f,")
                     append(" y1 = ${y1}f,")
                     append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f,")
-                    append(" x3 = ${x3}f,")
-                    append(" y3 = ${y3}f")
+                    append(" y2 = ${y2}f")
                     append(")")
-                }
+                },
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("curveToRelative(")
+                    append("reflectiveCurveToRelative(")
                     append("dx1 = ${x1}f,")
                     append(" dy1 = ${y1}f,")
                     append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f,")
-                    append(" dx3 = ${x3}f,")
-                    append(" dy3 = ${y3}f")
+                    append(" dy2 = ${y2}f")
                     append(")")
-                }
+                },
             )
         }
     }
 
     @Test
-    fun `ensure materialize generates curveTo with inlined parameters and no comment with close instruction when minified`() {
+    fun `ensure materialize generates reflectiveCurveTo with inlined parameters and no comment and close instruction when minified`() {
         // Arrange
-        val nonRelative = CurveToParams(
+        val nonRelative = ReflectiveCurveToParams(
             x1 = 15f,
             y1 = 12f,
             x2 = 10f,
             y2 = 5f,
-            x3 = 0f,
-            y3 = 123f,
         )
-        val relative = CurveToParams(
+        val relative = ReflectiveCurveToParams(
             x1 = 321f,
             y1 = 125f,
             x2 = 510f,
             y2 = 51f,
-            x3 = 10f,
-            y3 = 123f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "C${nonRelative}z c${relative}z"),
+            attributes = mutableMapOf("d" to "S${nonRelative}z s${relative}z"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -284,34 +246,30 @@ class PathNodesCurveToTests {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("curveTo(")
+                    append("reflectiveCurveTo(")
                     append("x1 = ${x1}f,")
                     append(" y1 = ${y1}f,")
                     append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f,")
-                    append(" x3 = ${x3}f,")
-                    append(" y3 = ${y3}f")
+                    append(" y2 = ${y2}f")
                     append(")")
                     appendLine()
                     append("close()")
-                }
+                },
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
                 element = buildString {
-                    append("curveToRelative(")
+                    append("reflectiveCurveToRelative(")
                     append("dx1 = ${x1}f,")
                     append(" dy1 = ${y1}f,")
                     append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f,")
-                    append(" dx3 = ${x3}f,")
-                    append(" dy3 = ${y3}f")
+                    append(" dy2 = ${y2}f")
                     append(")")
                     appendLine()
                     append("close()")
-                }
+                },
             )
         }
     }
