+++ shared/src/commonTest/kotlin/dev/tonholo/s2c/domain/svg/SvgPathReflectiveQuadToTests.kt
@@ -1,9 +1,7 @@
-package dev.tonholo.s2c.domain
+package dev.tonholo.s2c.domain.svg
 
-import dev.tonholo.s2c.domain.svg.SvgElementNode
-import dev.tonholo.s2c.domain.svg.SvgPathNode
-import dev.tonholo.s2c.domain.svg.asNode
-import dev.tonholo.s2c.domain.xml.XmlRootNode
+import dev.tonholo.s2c.domain.ImageVectorNode
+import dev.tonholo.s2c.domain.PathNodes
 import dev.tonholo.s2c.extensions.removeTrailingZero
 import kotlin.test.Test
 import kotlin.test.assertContains
@@ -12,28 +10,20 @@ import kotlin.test.assertFalse
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
-class PathNodesQuadToTests {
-    data class QuadToParams(
+class SvgPathReflectiveQuadToTests : BaseSvgTest() {
+    data class ReflectiveQuadToParams(
         val x1: Float,
         val y1: Float,
-        val x2: Float,
-        val y2: Float,
     ) {
-        override fun toString(): String = "$x1 $y1 $x2 $y2".removeTrailingZero()
+        override fun toString(): String = "$x1 $y1".removeTrailingZero()
     }
 
-    private val root = SvgElementNode(
-        parent = XmlRootNode(children = mutableSetOf()),
-        children = mutableSetOf(),
-        attributes = mutableMapOf(),
-    )
-
     @Test
-    fun `ensure a 'q' command is parsed to QuadTo relative`() {
+    fun `ensure a 't' command is parsed to ReflectiveQuadTo relative`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "q8,2 8,8"),
+            attributes = mutableMapOf("d" to "t8.5.2"), // should parse to t8.5 .2
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -41,17 +31,17 @@ class PathNodesQuadToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.QuadTo>(this)
+            assertIs<PathNodes.ReflectiveQuadTo>(this)
             assertTrue(isRelative)
         }
     }
 
     @Test
-    fun `ensure a 'S' command is parsed to QuadTo`() {
+    fun `ensure a 'T' command is parsed to ReflectiveQuadTo`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "Q2,-2 4,5"),
+            attributes = mutableMapOf("d" to "T -2.5-6.5"), // should parse to T-2.5 -6.5
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -59,29 +49,25 @@ class PathNodesQuadToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.QuadTo>(this)
+            assertIs<PathNodes.ReflectiveQuadTo>(this)
             assertFalse(isRelative)
         }
     }
 
     @Test
-    fun `ensure materialize generates quadTo instruction properly`() {
+    fun `ensure materialize generates reflectiveQuadTo instruction properly`() {
         // Arrange
-        val nonRelative = QuadToParams(
+        val nonRelative = ReflectiveQuadToParams(
             x1 = 15f,
             y1 = 12f,
-            x2 = 10f,
-            y2 = 5f,
         )
-        val relative = QuadToParams(
+        val relative = ReflectiveQuadToParams(
             x1 = 321f,
             y1 = 125f,
-            x2 = 510f,
-            y2 = 51f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "Q$nonRelative q$relative"),
+            attributes = mutableMapOf("d" to "T$nonRelative t$relative"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -92,12 +78,10 @@ class PathNodesQuadToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// Q $this
-                |quadTo(
+                |// T $this
+                |reflectiveQuadTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
-                |    x2 = ${x2}f,
-                |    y2 = ${y2}f,
                 |)
                 |""".trimMargin()
             )
@@ -106,12 +90,10 @@ class PathNodesQuadToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// q $this
-                |quadToRelative(
+                |// t $this
+                |reflectiveQuadToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
-                |    dx2 = ${x2}f,
-                |    dy2 = ${y2}f,
                 |)
                 |""".trimMargin()
             )
@@ -119,23 +101,19 @@ class PathNodesQuadToTests {
     }
 
     @Test
-    fun `ensure materialize generates quadTo with a close instruction properly`() {
+    fun `ensure materialize generates reflectiveQuadTo with a close instruction properly`() {
         // Arrange
-        val nonRelative = QuadToParams(
+        val nonRelative = ReflectiveQuadToParams(
             x1 = 15f,
             y1 = 12f,
-            x2 = 10f,
-            y2 = 5f,
         )
-        val relative = QuadToParams(
+        val relative = ReflectiveQuadToParams(
             x1 = 321f,
             y1 = 125f,
-            x2 = 510f,
-            y2 = 51f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "Q${nonRelative}z q${relative}z"),
+            attributes = mutableMapOf("d" to "T${nonRelative}z t${relative}z"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -146,12 +124,10 @@ class PathNodesQuadToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// Q ${this}z
-                |quadTo(
+                |// T ${this}z
+                |reflectiveQuadTo(
                 |    x1 = ${x1}f,
                 |    y1 = ${y1}f,
-                |    x2 = ${x2}f,
-                |    y2 = ${y2}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -161,12 +137,10 @@ class PathNodesQuadToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// q ${this}z
-                |quadToRelative(
+                |// t ${this}z
+                |reflectiveQuadToRelative(
                 |    dx1 = ${x1}f,
                 |    dy1 = ${y1}f,
-                |    dx2 = ${x2}f,
-                |    dy2 = ${y2}f,
                 |)
                 |close()
                 |""".trimMargin()
@@ -175,23 +149,19 @@ class PathNodesQuadToTests {
     }
 
     @Test
-    fun `ensure materialize generates quadTo with inlined parameters and no comment`() {
+    fun `ensure materialize generates reflectiveQuadTo with inlined parameters and no comment when minified`() {
         // Arrange
-        val nonRelative = QuadToParams(
+        val nonRelative = ReflectiveQuadToParams(
             x1 = 15f,
             y1 = 12f,
-            x2 = 10f,
-            y2 = 5f,
         )
-        val relative = QuadToParams(
+        val relative = ReflectiveQuadToParams(
             x1 = 321f,
             y1 = 125f,
-            x2 = 510f,
-            y2 = 51f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "Q$nonRelative q$relative"),
+            attributes = mutableMapOf("d" to "T$nonRelative t$relative"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -201,49 +171,31 @@ class PathNodesQuadToTests {
         with(nonRelative) {
             assertContains(
                 array = materialized,
-                element = buildString {
-                    append("quadTo(")
-                    append("x1 = ${x1}f,")
-                    append(" y1 = ${y1}f,")
-                    append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f")
-                    append(")")
-                },
+                element = "reflectiveQuadTo(x1 = ${x1}f, y1 = ${y1}f)",
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
-                element = buildString {
-                    append("quadToRelative(")
-                    append("dx1 = ${x1}f,")
-                    append(" dy1 = ${y1}f,")
-                    append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f")
-                    append(")")
-                },
+                element = "reflectiveQuadToRelative(dx1 = ${x1}f, dy1 = ${y1}f)",
             )
         }
     }
 
     @Test
-    fun `ensure materialize generates quadTo with inlined parameters and no comment with close instruction when minified`() {
+    fun `ensure materialize generates reflectiveQuadTo with inlined parameters and no comment and close instruction when minified`() {
         // Arrange
-        val nonRelative = QuadToParams(
+        val nonRelative = ReflectiveQuadToParams(
             x1 = 15f,
             y1 = 12f,
-            x2 = 10f,
-            y2 = 5f,
         )
-        val relative = QuadToParams(
+        val relative = ReflectiveQuadToParams(
             x1 = 321f,
             y1 = 125f,
-            x2 = 510f,
-            y2 = 51f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "Q${nonRelative}z q${relative}z"),
+            attributes = mutableMapOf("d" to "T${nonRelative}z t${relative}z"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -253,31 +205,19 @@ class PathNodesQuadToTests {
         with(nonRelative) {
             assertContains(
                 array = materialized,
-                element = buildString {
-                    append("quadTo(")
-                    append("x1 = ${x1}f,")
-                    append(" y1 = ${y1}f,")
-                    append(" x2 = ${x2}f,")
-                    append(" y2 = ${y2}f")
-                    append(")")
-                    appendLine()
-                    append("close()")
-                },
+                element = """
+                    |reflectiveQuadTo(x1 = ${x1}f, y1 = ${y1}f)
+                    |close()
+                """.trimMargin(),
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
-                element = buildString {
-                    append("quadToRelative(")
-                    append("dx1 = ${x1}f,")
-                    append(" dy1 = ${y1}f,")
-                    append(" dx2 = ${x2}f,")
-                    append(" dy2 = ${y2}f")
-                    append(")")
-                    appendLine()
-                    append("close()")
-                },
+                element = """
+                    |reflectiveQuadToRelative(dx1 = ${x1}f, dy1 = ${y1}f)
+                    |close()
+                """.trimMargin(),
             )
         }
     }
