+++ shared/src/commonTest/kotlin/dev/tonholo/s2c/domain/svg/SvgPathVerticalLineToTests.kt
@@ -1,9 +1,7 @@
-package dev.tonholo.s2c.domain
+package dev.tonholo.s2c.domain.svg
 
-import dev.tonholo.s2c.domain.svg.SvgElementNode
-import dev.tonholo.s2c.domain.svg.SvgPathNode
-import dev.tonholo.s2c.domain.svg.asNode
-import dev.tonholo.s2c.domain.xml.XmlRootNode
+import dev.tonholo.s2c.domain.ImageVectorNode
+import dev.tonholo.s2c.domain.PathNodes
 import dev.tonholo.s2c.extensions.removeTrailingZero
 import kotlin.test.Test
 import kotlin.test.assertContains
@@ -12,26 +10,19 @@ import kotlin.test.assertFalse
 import kotlin.test.assertIs
 import kotlin.test.assertTrue
 
-class PathNodesLineToTests {
-    data class LineParams(
-        val x: Float,
+class SvgPathVerticalLineToTests : BaseSvgTest() {
+    data class VerticalLineParams(
         val y: Float,
     ) {
-        override fun toString(): String = "$x $y".removeTrailingZero()
+        override fun toString(): String = "$y".removeTrailingZero()
     }
 
-    private val root = SvgElementNode(
-        parent = XmlRootNode(children = mutableSetOf()),
-        children = mutableSetOf(),
-        attributes = mutableMapOf(),
-    )
-
     @Test
-    fun `ensure a 'l' command is parsed to LineTo relative`() {
+    fun `ensure a 'v' command is parsed to VerticalLineTo relative`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "l 8 10"),
+            attributes = mutableMapOf("d" to "v 8"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -39,17 +30,17 @@ class PathNodesLineToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.LineTo>(this)
+            assertIs<PathNodes.VerticalLineTo>(this)
             assertTrue(isRelative)
         }
     }
 
     @Test
-    fun `ensure a 'L' command is parsed to LineTo`() {
+    fun `ensure a 'V' command is parsed to VerticalLineTo`() {
         // Arrange
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "L 8 10"),
+            attributes = mutableMapOf("d" to "V 8"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -57,25 +48,23 @@ class PathNodesLineToTests {
         // Assert
         assertEquals(expected = 1, actual = nodes.size)
         with(nodes.first()) {
-            assertIs<PathNodes.LineTo>(this)
+            assertIs<PathNodes.VerticalLineTo>(this)
             assertFalse(isRelative)
         }
     }
 
     @Test
-    fun `ensure materialize generates lineTo instruction properly`() {
+    fun `ensure materialize generates verticalLineTo instruction properly`() {
         // Arrange
-        val nonRelative = LineParams(
-            x = 8f,
-            y = 80f,
+        val nonRelative = VerticalLineParams(
+            y = 8f,
         )
-        val relative = LineParams(
-            x = 80f,
-            y = 10f,
+        val relative = VerticalLineParams(
+            y = 80f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "L$nonRelative l$relative"),
+            attributes = mutableMapOf("d" to "V$nonRelative v$relative"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -86,8 +75,8 @@ class PathNodesLineToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// L $this
-                |lineTo(x = ${x}f, y = ${y}f)
+                |// V $this
+                |verticalLineTo(y = ${y}f)
                 |""".trimMargin()
             )
         }
@@ -95,27 +84,25 @@ class PathNodesLineToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// l $this
-                |lineToRelative(dx = ${x}f, dy = ${y}f)
+                |// v $this
+                |verticalLineToRelative(dy = ${y}f)
                 |""".trimMargin()
             )
         }
     }
 
     @Test
-    fun `ensure materialize generates lineTo with a close instruction properly`() {
+    fun `ensure materialize generates verticalLineTo with a close instruction properly`() {
         // Arrange
-        val nonRelative = LineParams(
-            x = 8f,
-            y = 80f,
+        val nonRelative = VerticalLineParams(
+            y = 8f,
         )
-        val relative = LineParams(
-            x = 80f,
-            y = 10f,
+        val relative = VerticalLineParams(
+            y = 80f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "L${nonRelative}z l${relative}z"),
+            attributes = mutableMapOf("d" to "V${nonRelative}z v${relative}z"),
         )
         // Act
         val node = path.asNode() as ImageVectorNode.Path
@@ -126,8 +113,8 @@ class PathNodesLineToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// L ${this}z
-                |lineTo(x = ${x}f, y = ${y}f)
+                |// V ${this}z
+                |verticalLineTo(y = ${y}f)
                 |close()
                 |""".trimMargin()
             )
@@ -136,8 +123,8 @@ class PathNodesLineToTests {
             assertContains(
                 array = materialized,
                 element = """
-                |// l ${this}z
-                |lineToRelative(dx = ${x}f, dy = ${y}f)
+                |// v ${this}z
+                |verticalLineToRelative(dy = ${y}f)
                 |close()
                 |""".trimMargin()
             )
@@ -145,19 +132,17 @@ class PathNodesLineToTests {
     }
 
     @Test
-    fun `ensure materialize generates lineTo with no comment when minified`() {
+    fun `ensure materialize generates verticalLineTo with inlined parameters and no comment when minified`() {
         // Arrange
-        val nonRelative = LineParams(
-            x = 8f,
-            y = 80f,
+        val nonRelative = VerticalLineParams(
+            y = 8f,
         )
-        val relative = LineParams(
-            x = 80f,
-            y = 10f,
+        val relative = VerticalLineParams(
+            y = 80f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "L$nonRelative l$relative"),
+            attributes = mutableMapOf("d" to "V$nonRelative v$relative"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -167,31 +152,29 @@ class PathNodesLineToTests {
         with(nonRelative) {
             assertContains(
                 array = materialized,
-                element = "lineTo(x = ${x}f, y = ${y}f)"
+                element = "verticalLineTo(y = ${y}f)",
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
-                element = "lineToRelative(dx = ${x}f, dy = ${y}f)"
+                element = "verticalLineToRelative(dy = ${y}f)",
             )
         }
     }
 
     @Test
-    fun `ensure materialize generates lineTo with no comment and a close instruction when minified`() {
+    fun `ensure materialize generates verticalLineTo with inlined parameters and no comment and close instruction when minified`() {
         // Arrange
-        val nonRelative = LineParams(
-            x = 8f,
-            y = 80f,
+        val nonRelative = VerticalLineParams(
+            y = 8f,
         )
-        val relative = LineParams(
-            x = 80f,
-            y = 10f,
+        val relative = VerticalLineParams(
+            y = 80f,
         )
         val path = SvgPathNode(
             parent = root,
-            attributes = mutableMapOf("d" to "L${nonRelative}z l${relative}z"),
+            attributes = mutableMapOf("d" to "V${nonRelative}z v${relative}z"),
         )
         // Act
         val node = path.asNode(minified = true) as ImageVectorNode.Path
@@ -202,18 +185,18 @@ class PathNodesLineToTests {
             assertContains(
                 array = materialized,
                 element = """
-                    |lineTo(x = ${x}f, y = ${y}f)
-                    |close()
-                """.trimMargin()
+                |verticalLineTo(y = ${y}f)
+                |close()
+                """.trimMargin(),
             )
         }
         with(relative) {
             assertContains(
                 array = materialized,
                 element = """
-                    |lineToRelative(dx = ${x}f, dy = ${y}f)
-                    |close()
-                """.trimMargin()
+                |verticalLineToRelative(dy = ${y}f)
+                |close()
+                """.trimMargin(),
             )
         }
     }
